Fix recover calls, add ability to step into deferred functions.
The way `godebug.EnterFunc` is implemented was breaking calls to recover.
From the Go spec:

> The return value of recover is nil if any of the following conditions
> holds:
>   * panic's argument was nil;
>   * the goroutine is not panicking;
>   * recover was not called directly by a deferred function.

We pass the body of a function into `godebug.EnterFunc` so that it can
write information on the stack and then call the function as a callback
if it needs to. If the function calls recover, the recover call is no
longer directly inside any deferred function, so it always returns nil.

This commit fixes the problem by rewriting any calls to recover() as
channel receives on a channel we insert outside of the wrapped function.

See testdata/recover2-{in,out}.go for examples.