bug fix
For switch statements and select statements, godebug inserts extra cases
so the debugger can step through the evaluation of each (original) case.
If the switch/select was a terminating statement
(https://golang.org/ref/spec#Terminating_statements),
then these additions caused it to no longer be a terminating statement.
If it was the last statement in a function that returned a value, then
this caused the program to become invalid.

Simple solution: add `panic("impossible")` to the inserted cases.

I spent some time implementing a check to see if the switch/select was
terminating so I could avoid adding the panic if I didn't need it, since
it is ugly and unnecessary code much of the time. But when this check
grew to ~80-90 lines and I still hadn't implemented "has a break
statement referring to the enclosing for/switch/select" I decided it
wasn't worth it.