bug fix for type switch nested in expression switch
For expression switches, godebug generates case statements that let the
debugger pause between case statements in the original switch. For type
switches this isn't possible. As it is walking the AST, godebug tracks
whether it is inside an expression switch or not, so it knows whether to
output case statements when it encounters an ast.CaseClause. When it
encountered a type switch inside an expression switch, this bit was not
reset, so it proceeded to generate new case clauses inside the type
switch.

The fix is simple. The bad code was invalid Go, so just including an
instance of the nested switch statment in regression-in.go is a
sufficient test. I put the test code inside a function literal rather
than a function declaration because that was the context where I
originally found the bug.