warn user if an uninstrumented file has a breakpoint
I just had a confusing experience where I set a breakpoint in a
package in a subdirectory and then ran 'godebug run' on a file in the
working directory. godebug ran right past my breakpoint without
stopping. I added print statements to confirm that the program was
getting to that part of the code, and indeed it was.

I then realized my mistake: I had forgotten to pass the -instrument
flag. This situation is easy for godebug to catch, though, and the
breakpoint is useless if godebug is not instrumenting the package.
godebug should warn the user when this happens.