change breakpoint syntax
See #13

The current way to set a breakpoint in godebug is to call the function
`godebug.SetTrace()`. I don't like this for various reasons, the top one
being that it requires adding an import statement in addition to the
breakpoint line.

This commit changes the breakpoint marker to this:

    _ = "breakpoint"

This is only one line, is a no-op without godebug, and is unlikely to
conflict with users' program code.