do not allow variable names to be entered as commands
This removes the ambiguity where if I have a variable named "next" and I
type "next" into the console, it is not clear whether godebug should
print the variable or run the command by the same name. As of this
change, the only way to print the variable "next" is with "p next" (or
"print next").