bug fix for multiple-spec var declaration inside anonymous func
Resolves #47

godebug was producing incorrect code for variable declarations like
this:

    var (
        foo int
        bar int
    )

when these declarations were inside function literals. This is because
we were zeroing all of the token.Pos values in this syntax tree. Zeroing
these values is usually fine, but it turns out that go/printer assumes
that if the token for the left paren in a value spec is zero, then the
value spec only declares one value. Workaround: use token.Pos(1) instead
of token.Pos(0) for that paren.