bug fix: don't crash on method receivers named "_"
See #42

When godebug instruments a function, one of the things it adds requires
a reference to that function. For methods, this is usually achieved by
calling a function on the receiver:

    func (t MyType) Foo() {
        ...
        t.Foo()
        ...
    }

But methods do not have to have named receivers. godebug was already
correctly handling this case:

    func (MyType) Foo() {
        var receiver1 MyType
        ...
        receiver1.Foo()
        ...
    }

But godebug had failed to handle this very similar case, where the
receiver has an identifier, but the identifier is blank:

    func (_ MyType) Foo() {

The solution is the same as in the unnamed receiver case: Just create a
new variable of the same type and call the method on it.