feature: evaluate arbitrary expressions
Fixes #10 and fixes #18

Uses the library github.com/0xfaded/eval.

This commit also makes godebug aware of identifiers in the package
block, which fixes #11. This is accomplished by creating a package scope
object in the first file we encounter for a given package, and
initializing it with all of the identifiers in that package.