small cleanup of generated code
The packageFileTmpl template generated an extra newline inside the Vars,
Consts, and Funcs maps. Fixing this required moving the {{range}} Action
up to the previous line. I did not do this before because the previous
line ends in a literal "{" character, so the template would fail to
parse. But the function Delims, which I just discovered, lets you use a
different delimiter from "{{" and "}}". This lets me move the {{range}}
action up, delimit it as "<range>" to avoid the parse failure, and thus
clean up the extra newlines.